// packages/ui/src/types.ts
export interface Expense {
  id: number;
  eventId: number;
  label: string;
  amount: number;
  category: string;
  paid: boolean;
  receiptPath: string | null;
  date: string;
  notes: string | null;
  createdAt: string;
  updatedAt: string;
}